@@ -19,7 +19,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.4.1 <4.0.0"
+  sdk: ">=3.0.0-431.0.dev <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,37 +31,51 @@ dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  hive_generator: ^2.0.1
-  flutter_riverpod: ^2.5.1
-  equatable: ^2.0.5
-  build_runner: ^2.4.11
-  dio: ^5.4.3+1
-  awesome_dio_interceptor: ^1.2.0
+  uuid: ^3.0.7
+  path_provider: ^2.0.15
   dartz: ^0.10.1
-  uuid: ^4.4.0
-  image_picker: ^1.1.2
-  handle_wait_builder: ^0.9.0
-  permission_handler: ^11.3.1
-  path_provider: ^2.1.3
+  flutter_riverpod: ^2.3.6
+  dio: ^5.1.2
   pretty_dio_logger: ^1.3.1
-  shared_preferences: ^2.2.3
-  json_annotation: ^4.9.0
-  json_serializable: ^6.8.0
+  awesome_dio_interceptor: ^1.0.0
+  json_serializable: ^6.7.0
+  json_annotation: ^4.8.1
+  shared_preferences: ^2.1.1
+  connectivity_plus: ^4.0.1
+
+  cupertino_icons: ^1.0.2
+  multi_select_flutter: ^4.1.3
+  collection: ^1.17.1
+  image_picker: ^0.8.9
+  permission_handler: ^10.3.0
+  jwt_decoder: ^2.0.1
+  mockito: ^5.4.2
+  equatable: ^2.0.5
+  path: ^1.8.3
+  shake: ^2.2.0
+  all_sensors2: ^1.0.1
+  url_launcher: ^6.1.12
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  integration_test:
+    sdk: flutter
+  build_runner: ^2.4.4
+  hive_generator: ^2.0.0
+  hive_test: ^1.0.1
+  mocktail: ^0.3.0
+  test_cov_console: ^0.2.2
+  freezed_annotation: ^2.2.0
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.0
+  flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
