@@ -1,5 +1,5 @@
-import Flutter
 import UIKit
+import Flutter
 
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate {
