@@ -1,17 +1,19 @@
 import 'dart:io';
 
 import 'package:dartz/dartz.dart';
-import 'package:cshop/core/failure/failure.dart';
-import 'package:cshop/features/auth/data/repository/remote/auth_remote_repository.dart';
-import 'package:cshop/features/auth/domain/entity/auth_entity.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../../../../core/failure/failure.dart';
+import '../../data/repository/auth_remote_repository.dart';
+import '../entity/auth_entity.dart';
 
 final authRepositoryProvider = Provider<IAuthRepository>((ref) {
   return ref.read(authRemoteRepositoryProvider);
 });
 
 abstract class IAuthRepository {
-  Future<Either<Failure, bool>> registerStudent(AuthEntity student);
-  Future<Either<Failure, bool>> loginStudent(String username, String password);
+  Future<Either<Failure, bool>> registerUser(AuthEntity user);
+  Future<Either<Failure, bool>> loginUser(String username, String password);
   Future<Either<Failure, String>> uploadProfilePicture(File file);
+  
 }
