@@ -1,138 +1,209 @@
-import 'package:cshop/features/auth/presentation/viewmodel/auth_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../viewmodel/auth_view_model.dart';
+
 class LoginView extends ConsumerStatefulWidget {
-  const LoginView({super.key});
+  const LoginView({Key? key}) : super(key: key);
 
   @override
-  _LoginViewState createState() => _LoginViewState();
+  ConsumerState<LoginView> createState() => _LoginViewState();
 }
 
 class _LoginViewState extends ConsumerState<LoginView> {
   final _formKey = GlobalKey<FormState>();
   final _usernameController = TextEditingController();
   final _passwordController = TextEditingController();
+
   bool _obscurePassword = true;
+
+  void _togglePasswordVisibility() {
+    setState(() {
+      _obscurePassword = !_obscurePassword;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Center(
-          child: SingleChildScrollView(
-            padding: const EdgeInsets.all(16.0),
-            child: Form(
-              key: _formKey,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: <Widget>[
-                  const Text(
-                    'Poshak',
-                    textAlign: TextAlign.center,
-                    style: TextStyle(
-                      fontSize: 32,
-                      fontWeight: FontWeight.bold,
-                    ),
+    return SafeArea(
+      child: Scaffold(
+        backgroundColor: const Color.fromARGB(
+            255, 117, 196, 232), // Updated background color
+        body: SingleChildScrollView(
+          child: Padding(
+            padding: const EdgeInsets.all(20),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const SizedBox(height: 50),
+                const Text(
+                  'Trip Planners',
+                  style: TextStyle(
+                    color: Colors.black, // Updated text color
+                    fontWeight: FontWeight.bold,
+                    fontSize: 30.0,
                   ),
-                  const SizedBox(height: 40),
-                  TextFormField(
-                    key: const ValueKey('username'),
-                    controller: _usernameController,
-                    decoration: InputDecoration(
-                      labelText: 'Email',
-                      hintText: 'rajoli@gmail.com',
-                      prefixIcon: const Icon(Icons.account_circle),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                    ),
-                    validator: (value) {
-                      if (value!.isEmpty) {
-                        return 'Please enter email';
-                      } else if (!value.contains('@')) {
-                        return 'Email must contain @';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 20),
-                  TextFormField(
-                    key: const ValueKey('password'),
-                    controller: _passwordController,
-                    obscureText: _obscurePassword,
-                    decoration: InputDecoration(
-                      labelText: 'Password',
-                      prefixIcon: const Icon(Icons.lock),
-                      suffixIcon: IconButton(
-                        icon: Icon(
-                          _obscurePassword
-                              ? Icons.visibility_off
-                              : Icons.visibility,
+                ),
+                const SizedBox(height: 30),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    const Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          'Embark on Your Next Journey',
+                          style: TextStyle(
+                            color: Colors.black87, // Updated text color
+                            fontSize: 16.0,
+                          ),
                         ),
-                        onPressed: () {
-                          setState(() {
-                            _obscurePassword = !_obscurePassword;
-                          });
-                        },
-                      ),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12),
+                      ],
+                    ),
+                    const SizedBox(height: 20),
+                    Form(
+                      key: _formKey,
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          const Text(
+                            'Username',
+                            style: TextStyle(
+                              color: Colors.black, // Updated text color
+                              fontSize: 18.0,
+                              fontWeight: FontWeight.w600,
+                            ),
+                          ),
+                          const SizedBox(height: 8.0),
+                          TextFormField(
+                            key: const ValueKey('username'),
+                            controller: _usernameController,
+                            validator: (value) {
+                              if (value!.isEmpty) {
+                                return 'Please enter your username';
+                              }
+                              return null;
+                            },
+                            style: const TextStyle(
+                                color: Colors.black), // Updated text color
+                            decoration: InputDecoration(
+                              hintText: "Enter your username",
+                              hintStyle: TextStyle(
+                                  color: Colors
+                                      .grey[800]), // Updated hint text color
+                              border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(
+                                    10), // Updated border radius
+                              ),
+                              contentPadding: const EdgeInsets.all(16),
+                            ),
+                          ),
+                          const SizedBox(height: 20),
+                          const Text(
+                            'Password',
+                            style: TextStyle(
+                              color: Colors.black, // Updated text color
+                              fontSize: 18.0,
+                              fontWeight: FontWeight.w600,
+                            ),
+                          ),
+                          const SizedBox(height: 8.0),
+                          TextFormField(
+                            key: const ValueKey('password'),
+                            controller: _passwordController,
+                            validator: (value) {
+                              if (value!.isEmpty) {
+                                return 'Please enter your password';
+                              }
+                              return null;
+                            },
+                            obscureText: _obscurePassword,
+                            style: const TextStyle(
+                                color: Colors.black), // Updated text color
+                            decoration: InputDecoration(
+                              hintText: "Enter your password",
+                              hintStyle: TextStyle(
+                                  color: Colors
+                                      .grey[800]), // Updated hint text color
+                              border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(
+                                    10), // Updated border radius
+                              ),
+                              contentPadding: const EdgeInsets.all(16),
+                              suffixIcon: IconButton(
+                                icon: _obscurePassword
+                                    ? const Icon(Icons.visibility_off)
+                                    : const Icon(Icons.visibility),
+                                onPressed: _togglePasswordVisibility,
+                                color: Colors.black, // Updated icon color
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 15),
+                          SizedBox(
+                            width: double.infinity,
+                            height: 50,
+                            child: ElevatedButton(
+                              key: const ValueKey('loginButton'),
+                              onPressed: () async {
+                                if (_formKey.currentState!.validate()) {
+                                  await ref
+                                      .read(authViewModelProvider.notifier)
+                                      .loginUser(
+                                        context,
+                                        _usernameController.text.trim(),
+                                        _passwordController.text.trim(),
+                                      );
+                                }
+                              },
+                              style: ElevatedButton.styleFrom(
+                                shape: RoundedRectangleBorder(
+                                  borderRadius: BorderRadius.circular(
+                                      10), // Updated border radius
+                                ),
+                              ),
+                              child: const Text(
+                                'Sign In',
+                                style: TextStyle(
+                                  fontSize: 18.0,
+                                  fontWeight: FontWeight.w600,
+                                ),
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 15),
+                          const SizedBox(height: 5),
+                          Row(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            children: [
+                              const Text(
+                                "You don't have an account?",
+                                style: TextStyle(
+                                  color: Colors.black,
+                                  fontSize: 14,
+                                ),
+                              ),
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.pushNamed(context, "/register");
+                                },
+                                child: const Text(
+                                  'Sign up',
+                                  style: TextStyle(
+                                    color: Colors.black54,
+                                    fontSize: 14,
+                                    fontWeight: FontWeight.w600,
+                                  ),
+                                ),
+                              ),
+                            ],
+                          ),
+                        ],
                       ),
                     ),
-                    validator: (value) {
-                      if (value!.isEmpty) {
-                        return 'Please enter password';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 20),
-                  ElevatedButton(
-                    onPressed: () async {
-                      if (_formKey.currentState!.validate()) {
-                        await ref
-                            .read(authViewModelProvider.notifier)
-                            .loginStudent(
-                              _usernameController.text,
-                              _passwordController.text,
-                            );
-                      }
-                    },
-                    style: ElevatedButton.styleFrom(
-                      padding: const EdgeInsets.symmetric(vertical: 16),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                    ),
-                    child: const Text(
-                      'Login',
-                      style: TextStyle(
-                        fontSize: 18,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 16),
-                  TextButton(
-                    onPressed: () {
-                      ref
-                          .read(authViewModelProvider.notifier)
-                          .openRegisterView();
-                    },
-                    style: TextButton.styleFrom(
-                      padding: const EdgeInsets.all(16),
-                    ),
-                    child: const Text(
-                      'Don\'t have an account? Register',
-                      style: TextStyle(
-                        fontSize: 16,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
+                  ],
+                )
+              ],
             ),
           ),
         ),
