@@ -1,81 +1,82 @@
 import 'dart:io';
 
-import 'package:cshop/core/common/my_snackbar.dart';
-import 'package:cshop/features/auth/domain/entity/auth_entity.dart';
-import 'package:cshop/features/auth/domain/usecases/auth_usecase.dart';
-import 'package:cshop/features/auth/presentation/navigator/login_navigator.dart';
-import 'package:cshop/features/auth/presentation/state/auth_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final authViewModelProvider = StateNotifierProvider<AuthViewModel, AuthState>(
-  (ref) => AuthViewModel(
-    ref.read(loginViewNavigatorProvider),
+
+import 'package:tripplanner/config/router/app_route.dart';
+import 'package:tripplanner/core/common/snackbar/my_snackbar.dart';
+
+import '../../domain/entity/auth_entity.dart';
+import '../../domain/use_case/auth_usecase.dart';
+import '../state/auth_state.dart';
+
+final authViewModelProvider =
+    StateNotifierProvider<AuthViewModel, AuthState>((ref) {
+  return AuthViewModel(
     ref.read(authUseCaseProvider),
-  ),
-);
+  );
+});
 
 class AuthViewModel extends StateNotifier<AuthState> {
-  AuthViewModel(this.navigator, this.authUseCase) : super(AuthState.initial());
-  final AuthUseCase authUseCase;
-  final LoginViewNavigator navigator;
+  final AuthUseCase _authUseCase;
 
-  Future<void> uploadImage(File? file) async {
-    state = state.copyWith(isLoading: true);
-    var data = await authUseCase.uploadProfilePicture(file!);
-    data.fold(
-      (l) {
-        state = state.copyWith(isLoading: false, error: l.error);
-        showMySnackBar(message: l.error);
-      },
-      (imageName) {
-        state =
-            state.copyWith(isLoading: false, error: null, imageName: imageName);
-      },
-    );
-  }
+  AuthViewModel(this._authUseCase) : super(AuthState(isLoading: false));
 
-  Future<void> registerStudent(AuthEntity student) async {
+  Future<void> registerUser(AuthEntity user,BuildContext context) async {
+    
     state = state.copyWith(isLoading: true);
-    var data = await authUseCase.registerStudent(student);
-    data.fold(
+    var data = await _authUseCase.registerUser(user);
+     data.fold(
       (failure) {
-        state = state.copyWith(
-          isLoading: false,
-          error: failure.error,
-        );
-        showMySnackBar(message: failure.error, color: Colors.red);
+        state = state.copyWith(isLoading: false, error: failure.error);
+        showSnackBar(
+            message: "Signup Failed", context: context, color: Colors.red);
       },
       (success) {
         state = state.copyWith(isLoading: false, error: null);
-        showMySnackBar(message: "Successfully registered");
+        showSnackBar(
+            message: 'Signup Success', context: context, color: Colors.green);
+        Navigator.popAndPushNamed(context, AppRoute.homeRoute);
       },
     );
   }
 
-  Future<void> loginStudent(
+  Future<void> loginUser(
+    BuildContext context,
     String username,
     String password,
   ) async {
     state = state.copyWith(isLoading: true);
-    var data = await authUseCase.loginStudent(username, password);
+    var data = await _authUseCase.loginUser(username, password);
     data.fold(
       (failure) {
         state = state.copyWith(isLoading: false, error: failure.error);
-        showMySnackBar(message: failure.error, color: Colors.red);
+        showSnackBar(
+            message: "Login Failed", context: context, color: Colors.red);
       },
       (success) {
         state = state.copyWith(isLoading: false, error: null);
-        openHomeView();
+        showSnackBar(
+            message: 'Login Success', context: context, color: Colors.green);
+        Navigator.popAndPushNamed(context, AppRoute.homeRoute);
       },
     );
   }
 
-  void openRegisterView() {
-    navigator.openRegisterView();
-  }
-
-  void openHomeView() {
-    navigator.openHomeView();
+  Future<void> uploadImage(File file) async {
+    state = state.copyWith(isLoading: true);
+    var data = await _authUseCase.uploadProfilePicture(file);
+    data.fold(
+        (failure) => state = state.copyWith(
+              isLoading: false,
+              error: failure.error,
+            ), (imageName) {
+      state = state.copyWith(
+        isLoading: false,
+        error: null,
+        imageName: imageName,
+      );
+    });
   }
 }
