@@ -1,6 +1,7 @@
-import 'package:cshop/features/splash/presentation/viewmodel/splash_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../viewmodel/splash_view_model.dart';
 
 class SplashView extends ConsumerStatefulWidget {
   const SplashView({super.key});
@@ -12,55 +13,54 @@ class SplashView extends ConsumerStatefulWidget {
 class _SplashViewState extends ConsumerState<SplashView> {
   @override
   void initState() {
-    ref.read(splashViewModelProvider.notifier).openLoginView();
-
     // Wait for 2 seconds and then navigate
-    // Future.delayed(const Duration(seconds: 2), () {
-    //   Navigator.of(context).pushReplacement(MaterialPageRoute(
-    //     builder: (context) => const LoginView(),
-    //   ));
-    // });
+    Future.delayed(const Duration(seconds: 2), () {
+      ref.read(splashViewModelProvider.notifier).init(context);
+    });
 
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.green[50],
-      body: Stack(
-        children: [
-          Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                SizedBox(
-                  height: 200,
-                  width: 200,
-                  child: Image.asset('assets/images/image17.jpg'),
-                ),
-                const Text(
-                  'Digitized clothing app',
-                  style: TextStyle(
-                    fontSize: 25,
-                  ),
-                ),
-                const SizedBox(height: 10),
-                const CircularProgressIndicator(),
-                const SizedBox(height: 10),
-                const Text('version : 1.0.0')
-              ],
+    return const Scaffold(
+      backgroundColor:
+          Colors.blueAccent, // Updated background color for Trip Planner
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Trip Planner',
+              style: TextStyle(
+                fontSize: 36,
+                fontWeight: FontWeight.bold,
+                color: Colors.white,
+              ),
             ),
-          ),
-          Positioned(
-            bottom: 10,
-            left: MediaQuery.of(context).size.width / 4,
-            child: const Text(
-              'Developed by: Raj Oli',
-              style: TextStyle(fontSize: 15),
+            SizedBox(height: 20),
+            Text(
+              'Plan Your Next Adventure',
+              style: TextStyle(
+                fontSize: 18,
+                color: Colors.white70,
+              ),
+              textAlign: TextAlign.center,
             ),
-          ),
-        ],
+            SizedBox(height: 20),
+            CircularProgressIndicator(
+              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+            ),
+            SizedBox(height: 20),
+            Text(
+              'Created By: Rabin Shahi',
+              style: TextStyle(
+                fontSize: 18,
+                color: Colors.white,
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
