@@ -1,25 +1,49 @@
-import 'package:cshop/features/splash/presentation/navigator/splash_navigator.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:jwt_decoder/jwt_decoder.dart';
 
-final splashViewModelProvider =
-    StateNotifierProvider<SplashViewModel, void>((ref) {
-  final navigator = ref.read(splashViewNavigatorProvider);
-  return SplashViewModel(navigator);
-});
+import '../../../../config/router/app_route.dart';
+import '../../../../core/shared_prefs/user_shared_prefs.dart';
+
+final splashViewModelProvider = StateNotifierProvider<SplashViewModel, void>(
+  (ref) {
+    return SplashViewModel(
+      ref.read(userSharedPrefsProvider),
+    );
+  },
+);
 
 class SplashViewModel extends StateNotifier<void> {
-  SplashViewModel(this.navigator) : super(null);
+  final UserSharedPrefs _userSharedPrefs;
+  SplashViewModel(this._userSharedPrefs) : super(null);
 
-  final SplashViewNavigator navigator;
+  init(BuildContext context) async {
+    final data = await _userSharedPrefs.getUserToken();
 
-  //open login page
-  void openLoginView() {
-    Future.delayed(const Duration(seconds: 2), () {
-      navigator.openLoginView();
+    data.fold((l) => null, (token) {
+      if (token != null && token.isNotEmpty) {
+        bool isTokenExpired = isValidToken(token);
+        if (isTokenExpired) {
+          Navigator.popAndPushNamed(context, AppRoute.loginRoute);
+        } else {
+          Navigator.popAndPushNamed(context, AppRoute.homeRoute);
+        }
+      } else {
+        Navigator.popAndPushNamed(context, AppRoute.loginRoute);
+      }
     });
   }
 
-  //later on we will add open home page method here as well
-
-  void openHomeView() {}
+  bool isValidToken(String token) {
+    try {
+      Map<String, dynamic> decodedToken = JwtDecoder.decode(token);
+      int expirationTimestamp = decodedToken['exp'];
+      final currentDate = DateTime.now().millisecondsSinceEpoch;
+      // If the current date is greater than the expiration timestamp, the token is expired
+      return currentDate > expirationTimestamp * 1000;
+    } catch (e) {
+      print('Error decoding token: $e');
+      return false;
+    }
+  }
 }
